+++ flutter/realm_flutter/example/lib/main.dart
@@ -1,9 +1,27 @@
+////////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2021 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////////
+
 import 'package:flutter/material.dart';
 import 'dart:async';
 
 import 'package:flutter/services.dart';
-// import 'package:realm_flutter/realm_flutter.dart';
-import 'package:realm_flutter/realm.dart';
+
+import 'package:realm/realm.dart';
 
 part 'main.g.dart';
 
@@ -13,6 +31,7 @@ class _Car {
 }
 
 void main() {
+  initRealm();
   runApp(MyApp());
 }