+++ flutter/realm_flutter/ios/Classes/RealmFlutterPlugin.m
@@ -1,15 +1,21 @@
 #import "RealmFlutterPlugin.h"
-#if __has_include(<realm_flutter/realm_flutter-Swift.h>)
-#import <realm_flutter/realm_flutter-Swift.h>
-#else
-// Support project import fallback if the generated compatibility header
-// is not copied when this plugin is created as a library.
-// https://forums.swift.org/t/swift-static-libraries-dont-copy-generated-objective-c-header/19816
-#import "realm_flutter-Swift.h"
-#endif
 
 @implementation RealmFlutterPlugin
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  [SwiftRealmFlutterPlugin registerWithRegistrar:registrar];
+  FlutterMethodChannel* channel = [FlutterMethodChannel
+      methodChannelWithName:@"realm_flutter"
+            binaryMessenger:[registrar messenger]];
+  RealmFlutterPlugin* instance = [[RealmFlutterPlugin alloc] init];
+  [registrar addMethodCallDelegate:instance channel:channel];
 }
+
+- (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
+  if ([@"getPlatformVersion" isEqualToString:call.method]) {
+    result([@"iOS " stringByAppendingString:[[UIDevice currentDevice] systemVersion]]);
+  } else {
+    result(FlutterMethodNotImplemented);
+  }
+}
+
 @end
+