+++ lib/src/list.dart
@@ -1,9 +1,29 @@
+////////////////////////////////////////////////////////////////////////////////
+//
+// Copyright 2021 Realm Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////////
+
 import 'dart:core';
 import 'dart:core' as core;
 import 'dart:collection';
 
 import 'realm_object.dart';
 
+
+/// @nodoc
 class RealmList {
   /**
    * Called from native code
@@ -105,9 +125,10 @@ class RealmList {
  * Dart: Have two classes implement the same interface and swap the unamanged and managed 
  * implementation instead of checking _unmanagedValues in every member
  */
+/// @nodoc
 class _RealmListUnmanaged {}
 
-
+/// A list of RealmObjects
 class ArrayList<T extends RealmObject> with ListMixin<T> {
   RealmList _list;
   core.List<T> _unmanagedValues;