+++ packages/realm/android/src/main/cpp/lib/include/realm/string_data.hpp
@@ -1,0 +1,1 @@
+../../../../../../../../../flutter/realm_flutter/android/src/main/cpp/lib/include/realm/string_data.hpp