+++ packages/realm/ios/src/dart-include/dart_native_api.h
@@ -1,0 +1,1 @@
+../../../../../flutter/realm_flutter/ios/src/dart-include/dart_native_api.h