+++ packages/realm/ios/src/realm-dart-extension/realm_dart_extension.cpp
@@ -1,0 +1,1 @@
+../../../../../flutter/realm_flutter/ios/src/realm-dart-extension/realm_dart_extension.cpp