+++ packages/realm_dart/src/object-store/src/binding_callback_thread_observer.hpp
@@ -1,0 +1,1 @@
+../../../../../src/object-store/src/binding_callback_thread_observer.hpp