+++ packages/realm_dart/src/object-store/src/collection_notifications.hpp
@@ -1,0 +1,1 @@
+../../../../../src/object-store/src/collection_notifications.hpp