+++ packages/realm_dart/src/object-store/src/shared_realm.hpp
@@ -1,0 +1,1 @@
+../../../../../src/object-store/src/shared_realm.hpp