+++ packages/realm_dart/src/object-store/src/sync/sync_user.cpp
@@ -1,0 +1,1 @@
+../../../../../../src/object-store/src/sync/sync_user.cpp