+++ packages/realm_dart/src/vendor-include/realm-win-x64/include/realm/array_string_short.hpp
@@ -1,0 +1,1 @@
+../../../../../../../src/vendor-include/realm-win-x64/include/realm/array_string_short.hpp