+++ flutter/realm_flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.1
+0.1.0+preview Release notes (2021-04-01)
+=============================================================
+### Enhancements
+* The initial preview version of the Realm SDK for Flutter.
 
-* TODO: Describe initial release.
+### Compatibility
+* Flutter 2.0
+* Android Emulators and Android Devices
+* iOS simulator